added app button button panel display as per code review
