--- src/App.js
@@ -1,15 +1,0 @@
-import './App.css';
-import React from 'react';
-import Calculator from './calculator';
-
-function App() {
-  return (
-    <div className="App">
-      <Calculator />
-    </div>
-  );
-}
-
-export default App;
-
-// Render the Calculator to the Web page.
