--- src/calculatorTitle.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-// Create a Class Component Calculator.
-class Calculator extends React.Component {
-  render() {
-    return 'Calculator';
-  }
-}
-
-// Export Calculator.
-export default Calculator;
