--- src/component.module.css
@@ -1,11 +1,0 @@
-.buton {
-  padding: 5px;
-  width: 5vw;
-  height: 5vh;
-  font-size: larger;
-}
-
-input {
-  width: 20%;
-  height: 50px;
-}
