+++ src/component/App/App.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import styles from './App.module.css';
+import ButtonPanel from '../buttonPanel';
+
+function App() {
+  return (
+    <div className={styles.App}>
+      <ButtonPanel />
+    </div>
+  );
+}
+
+export default App;
