+++ src/component/Display/index.js
@@ -1,0 +1,3 @@
+import Display from './Display';
+
+export default Display;
