+++ src/component/button/button.module.css
@@ -1,0 +1,5 @@
+.button {
+  width: 100%;
+  height: 40px;
+  border: 1px solid #000;
+}
