+++ src/component/button/index.js
@@ -1,0 +1,3 @@
+import Button from './button';
+
+export default Button;
