+++ src/component/buttonPanel/index.js
@@ -1,0 +1,3 @@
+import ButtonPanel from './buttonPanel';
+
+export default ButtonPanel;
